Move galaxy.tools.toolbox.cache to galaxy.tools.cache.
galaxy.tools.toolbox is shared with galaxy-lib and so should have no dependencies on sqlalchemy. Also this really seems to be caching tools and tool shed repositories and fairly independent of the toolbox abstraction so I think it is solid refactoring independent of the sqlalchemy/galaxy-lib constraints.
